Staging

+++ src/pageTemplates/InventoryTemplate/index.tsx
@@ -1,32 +1,38 @@
-import { LayoutWithSidebar } from "@/components/layouts/LayoutWithSidebar";
-import { Heading } from "@/components/Heading";
-import { Table } from "@/components/Table";
-import { Paragraph } from "@/components/Paragraph";
-import { Button } from "@/components/Button";
-import { useCallback, useEffect, useState } from "react";
-import { PlusCircle } from "phosphor-react";
-import { ModalAddInventory } from "../../components/layouts/modals/ModalAddInventory";
-import { ModalEditInventory } from "@/components/layouts/modals/ModalEditInventory";
+import { LayoutWithSidebar } from '@/components/layouts/LayoutWithSidebar';
+import { Heading } from '@/components/Heading';
+import { Table } from '@/components/Table';
+import { Paragraph } from '@/components/Paragraph';
+import { Button } from '@/components/Button';
+import { useCallback, useEffect, useState } from 'react';
+import { PlusCircle } from 'phosphor-react';
+import { ModalAddInventory } from '../../components/layouts/modals/ModalAddInventory';
+import { ModalEditInventory } from '@/components/layouts/modals/ModalEditInventory';
 import {
   addInventory,
   deleteInventory,
   fetchInventory,
   editInventory,
-} from "@/api/inventory";
-import { handleToast } from "@/utils/handleToast";
-import { formatCurrencyToFloat } from "@/utils/formatCurrencyToFloat";
+} from '@/api/inventory';
+import { handleToast } from '@/utils/handleToast';
+import { formatCurrencyToFloat } from '@/utils/formatCurrencyToFloat';
 import {
   IAddInventoryBody,
   IEditInventoryBody,
   IInventoryModel,
-} from "@/@types/inventory";
+} from '@/@types/inventory';
 
-import { formatDateToDDMMYYYY } from "@/utils/formatDateToDDMMYYYY";
-import { convertRealToQuantity } from "@/utils/convertRealToQuantity";
-import { formatDDMMYYYYToDate } from "@/utils/formatDDMMYYYYToDate";
+import { formatDateToDDMMYYYY } from '@/utils/formatDateToDDMMYYYY';
+import { convertRealToQuantity } from '@/utils/convertRealToQuantity';
+import { formatDDMMYYYYToDate } from '@/utils/formatDDMMYYYYToDate';
+import { sortItems } from '@/utils/sortItems';
+import Swal from 'sweetalert2';
+import { useAuth } from '@/hooks/useAuth';
+
+const ITEMS_SORT = ['validity', 'quantity'];
 
 export const InventoryTemplate = () => {
   const [inventory, setInventory] = useState([]);
+  const [sortDirection, setSortDirection] = useState<'ASC' | 'DESC'>('DESC');
   const [modalAddInventoryIsOpen, setModalAddInventoryIsOpen] = useState(false);
   const [modalEditInventoryIsOpen, setModalEditInventoryIsOpen] =
     useState(false);
@@ -36,12 +42,14 @@ export const InventoryTemplate = () => {
     {} as IInventoryModel
   );
 
+  const { isAuthenticated } = useAuth();
+
   const handleFetchInventory = useCallback(
     async (search?: string, page?: number) => {
       setIsLoading(true);
       try {
         const { inventory, totalPages } = await fetchInventory(
-          search ?? "",
+          search ?? '',
           page ?? 1
         );
         setInventory(inventory);
@@ -64,7 +72,6 @@ export const InventoryTemplate = () => {
   const handleEditInventory = async (values: IEditInventoryBody) => {
     setIsLoading(true);
     try {
-      console.log("==> ", formatDDMMYYYYToDate(values.validity).toString());
       await editInventory(
         {
           ...values,
@@ -73,13 +80,13 @@ export const InventoryTemplate = () => {
         currentInventory.id
       );
       await handleFetchInventory();
-      handleToast("Estoque editado com sucesso.", "success");
+      handleToast('Estoque editado com sucesso.', 'success');
     } catch (err) {
       if (err.response.data.err) {
-        handleToast(err.response.data.err, "error");
+        handleToast(err.response.data.err, 'error');
         return;
       }
-      handleToast("Erro ao editar membro.", "error");
+      handleToast('Erro ao editar membro.', 'error');
     } finally {
       setIsLoading(false);
       setModalEditInventoryIsOpen(false);
@@ -102,32 +109,62 @@ export const InventoryTemplate = () => {
         quantity,
         validity: formatDateToDDMMYYYY(validity),
       });
-      handleToast("Estoque adicionado com sucesso.", "success");
+      handleToast('Estoque adicionado com sucesso.', 'success');
       setModalAddInventoryIsOpen(false);
       handleFetchInventory();
     } catch (err) {
-      handleToast("Algo aconteceu de errado.", "error");
+      handleToast('Algo aconteceu de errado.', 'error');
     } finally {
       setIsLoading(false);
     }
   };
 
   const handleDeleteItem = async (inventoryId: string) => {
-    setIsLoading(true);
     try {
-      await deleteInventory(inventoryId);
+      Swal.fire({
+        title: 'Você tem certeza?',
+        text: 'Essa ação é irreversível!',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Sim, deletar!',
+        cancelButtonText: 'Cancelar',
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          setIsLoading(true);
+          await deleteInventory(inventoryId);
+          handleToast('Estoque deletado com sucesso.', 'success');
+          handleFetchInventory();
+          setIsLoading(false);
+        }
+      });
     } catch (err) {
-      console.log(err);
+      console.error('Erro ao exibir pop-up de confirmação:', err);
     } finally {
-      handleFetchInventory();
       setIsLoading(false);
-      handleToast("Estoque deletado com sucesso.", "success");
     }
+  };
+
+  const handleSortItems = (title: string) => {
+    let items = [];
+
+    switch (title) {
+      case 'quantity':
+        items = sortItems(inventory, 'quantity', sortDirection, 'number');
+
+      case 'validity':
+        items = sortItems(inventory, 'validity', sortDirection, 'date');
+    }
+
+    setSortDirection(sortDirection === 'ASC' ? 'DESC' : 'ASC');
+
+    setInventory(items);
   };
 
   useEffect(() => {
     handleFetchInventory();
-  }, [handleFetchInventory]);
+  }, [handleFetchInventory, isAuthenticated]);
 
   return (
     <>
@@ -154,6 +191,8 @@ export const InventoryTemplate = () => {
             isLoading={isLoading}
             handleGetItemsWithSearch={handleFetchInventory}
             totalPage={totalPages}
+            itemsSort={ITEMS_SORT}
+            handleSortItems={handleSortItems}
           />
         </div>
       </LayoutWithSidebar>
