+++ src/pageTemplates/ProductsTemplate/index.tsx
@@ -1,23 +1,23 @@
-import { IEditProduct, IProductModel } from "@/@types/product";
-import { IEditMember } from "@/@types/user";
+import { IEditProduct, IProductModel } from '@/@types/product';
 import {
   addProduct,
   deleteProduct,
   editProduct,
   fetchProducts,
   IAddProductBody,
-} from "@/api/products";
-import { editUser } from "@/api/user";
-import { Button } from "@/components/Button";
-import { Heading } from "@/components/Heading";
-import { LayoutWithSidebar } from "@/components/layouts/LayoutWithSidebar";
-import { ModalAddProduct } from "@/components/layouts/modals/ModalAddProduct";
-import { ModalEditProduct } from "@/components/layouts/modals/ModalEditProduct";
-import { Paragraph } from "@/components/Paragraph";
-import { Table } from "@/components/Table";
-import { handleToast } from "@/utils/handleToast";
-import { PlusCircle } from "phosphor-react";
-import { useCallback, useEffect, useState } from "react";
+} from '@/api/products';
+import { Button } from '@/components/Button';
+import { Heading } from '@/components/Heading';
+import { LayoutWithSidebar } from '@/components/layouts/LayoutWithSidebar';
+import { ModalAddProduct } from '@/components/layouts/modals/ModalAddProduct';
+import { ModalEditProduct } from '@/components/layouts/modals/ModalEditProduct';
+import { Paragraph } from '@/components/Paragraph';
+import { Table } from '@/components/Table';
+import { useAuth } from '@/hooks/useAuth';
+import { handleToast } from '@/utils/handleToast';
+import { PlusCircle } from 'phosphor-react';
+import { useCallback, useEffect, useState } from 'react';
+import Swal from 'sweetalert2';
 
 export const ProductsTemplate = () => {
   const [modalAddProductIsOpen, setModalAddProductIsOpen] = useState(false);
@@ -26,6 +26,8 @@ export const ProductsTemplate = () => {
   const [products, setProducts] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
   const [totalPages, setTotalPages] = useState(1);
+
+  const { isAuthenticated } = useAuth();
 
   const handleOpenEditProduct = (productId: string) => {
     const item = products.find((product) => product.id == productId);
@@ -38,13 +40,13 @@ export const ProductsTemplate = () => {
     try {
       await editProduct(values, currentProduct.id);
       await handleFetchProducts();
-      handleToast("Produto editado com sucesso.", "success");
+      handleToast('Produto editado com sucesso.', 'success');
     } catch (err) {
       if (err.response.data.err) {
-        handleToast(err.response.data.err, "error");
+        handleToast(err.response.data.err, 'error');
         return;
       }
-      handleToast("Erro ao editar produto.", "error");
+      handleToast('Erro ao editar produto.', 'error');
     } finally {
       setIsLoading(false);
       setModalEditProductIsOpen(false);
@@ -52,15 +54,29 @@ export const ProductsTemplate = () => {
   };
 
   const handleDeleteItem = async (productId: string) => {
-    setIsLoading(true);
     try {
-      await deleteProduct(productId);
+      Swal.fire({
+        title: 'Você tem certeza?',
+        text: 'Essa ação é irreversível!',
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Sim, deletar!',
+        cancelButtonText: 'Cancelar',
+      }).then(async (result) => {
+        if (result.isConfirmed) {
+          setIsLoading(true);
+          await deleteProduct(productId);
+          handleToast('Produto deletado com sucesso.', 'success');
+          setIsLoading(false);
+          handleFetchProducts();
+        }
+      });
     } catch (err) {
       console.log(err);
     } finally {
-      handleFetchProducts();
       setIsLoading(false);
-      handleToast("Produto deletado com sucesso.", "success");
     }
   };
 
@@ -68,10 +84,10 @@ export const ProductsTemplate = () => {
     setIsLoading(true);
     try {
       await addProduct({ name, sigla });
-      handleToast("Produto adicionado com sucesso.", "success");
+      handleToast('Produto adicionado com sucesso.', 'success');
       setModalAddProductIsOpen(false);
     } catch (err) {
-      handleToast("Erro ao adicionar produto.", "error");
+      handleToast('Erro ao adicionar produto.', 'error');
     } finally {
       setIsLoading(false);
       handleFetchProducts();
@@ -83,7 +99,7 @@ export const ProductsTemplate = () => {
       setIsLoading(true);
       try {
         const { products, totalPages } = await fetchProducts(
-          search ?? "",
+          search ?? '',
           page ?? 1
         );
         setProducts(products);
@@ -99,7 +115,7 @@ export const ProductsTemplate = () => {
 
   useEffect(() => {
     handleFetchProducts();
-  }, []);
+  }, [handleFetchProducts, isAuthenticated]);
 
   return (
     <>
@@ -118,6 +134,7 @@ export const ProductsTemplate = () => {
               Adicionar produto
             </Button>
           </div>
+
           <Table
             content={products}
             handleEditItem={handleOpenEditProduct}
